+++ packages/studio-core/src/lib/hooks/index.ts
@@ -1,3 +1,2 @@
 export * from './useT';
-export * from './useLiquidGlassNav';
 export * from './navScroll';